Refactor get_downloadable_from_file into File.as_downloadable
